Move WebHookHandler to handlers module
