+++ src/services/webHookHandler.ts
@@ -1,4 +1,4 @@
-import { WebHookHandler } from "../WebHookHandler";
+import { WebHookHandler } from "../handlers/WebHookHandler";
 import logger from "./logger";
 import queue from "./queue";
